[DOM-255] Refactor OfferDetails to a separate flow
