+++ Domostroy/Domostroy/Flows/Favorites/FavoritesCoordinator.swift
@@ -35,9 +35,15 @@ private extension FavoritesCoordinator {
     func showFavorites() {
         let (view, output) = FavoritesModuleConfigurator().configure()
         output.onOpenOffer = { [weak self] id in
-
+            self?.runOfferDetailsFlow(id: id)
         }
         router.setNavigationControllerRootModule(view, animated: false, hideBar: false)
+    }
+
+    func runOfferDetailsFlow(id: Int) {
+        let coordinator = OfferDetailsCoordinator(router: router)
+        addDependency(coordinator)
+        coordinator.start(with: id)
     }
 
 }
