+++ Domostroy/Domostroy/Flows/OfferDetails/OfferDetailsCoordinatorOutput.swift
@@ -1,0 +1,13 @@
+//
+//  OfferDetailsCoordinatorOutput.swift
+//  Domostroy
+//
+//  Created by igorpustylnik on 03/04/2025.
+//  Copyright © 2025 Domostroy. All rights reserved.
+//
+
+import UIKit
+
+protocol OfferDetailsCoordinatorOutput: AnyObject {
+
+}
